Updating GetUser to use string
